remove internal copy of swipable state and create new PreUpPostDownNestedScrollConnection object
